requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Web Application
streamlit>=1.28.0
plotly>=5.17.0

# Model serving (optional)
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Jupyter notebooks (optional)
# jupyter>=1.0.0
# ipywidgets>=8.1.0

# Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0